refactored api calls into 1 component
